requirements.txt
cffi==2.0.0
cryptography==46.0.7
customtkinter==5.2.2
darkdetect==0.8.0
packaging==26.1
Pillow>=10.0.0
pycparser==3.0
